fix(refactoring): refactored domain objects
