+++ pom.xml
@@ -437,6 +437,8 @@
                     <connectionUrl>scm:git:https://github.com/pagopa/pagopa-ecommerce-commons.git</connectionUrl>
                     <scmVersionType>tag</scmVersionType>
                     <scmVersion>${pagopa-ecommerce-commons.version}</scmVersion>
+                    <scmVersionType>branch</scmVersionType>
+                    <scmVersion>CHK-1921-domain-object-unversioning</scmVersion>
                     <goals>install -DskipTests</goals>
                 </configuration>
                 <executions>
