+++ src/main/java/it/pagopa/transactions/commands/TransactionActivateCommand.java
@@ -1,8 +1,8 @@
 package it.pagopa.transactions.commands;
 
 import it.pagopa.ecommerce.commons.documents.v1.Transaction;
-import it.pagopa.ecommerce.commons.domain.v1.RptId;
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId;
+import it.pagopa.ecommerce.commons.domain.RptId;
+import it.pagopa.ecommerce.commons.domain.TransactionId;
 import it.pagopa.generated.transactions.server.model.NewTransactionRequestDto;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
