+++ src/main/java/it/pagopa/transactions/commands/TransactionClosureSendCommand.java
@@ -1,6 +1,6 @@
 package it.pagopa.transactions.commands;
 
-import it.pagopa.ecommerce.commons.domain.v1.*;
+import it.pagopa.ecommerce.commons.domain.RptId;
 import it.pagopa.transactions.commands.data.ClosureSendData;
 
 public final class TransactionClosureSendCommand extends TransactionsCommand<ClosureSendData> {
