+++ src/main/java/it/pagopa/transactions/mdcutilities/TransactionTracingUtils.java
@@ -1,7 +1,7 @@
 package it.pagopa.transactions.mdcutilities;
 
-import it.pagopa.ecommerce.commons.domain.v1.RptId;
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId;
+import it.pagopa.ecommerce.commons.domain.RptId;
+import it.pagopa.ecommerce.commons.domain.TransactionId;
 import reactor.util.context.Context;
 
 import java.util.Set;
