+++ src/main/java/it/pagopa/transactions/projections/TransactionsProjection.java
@@ -1,6 +1,6 @@
 package it.pagopa.transactions.projections;
 
-import it.pagopa.ecommerce.commons.domain.v1.*;
+import it.pagopa.ecommerce.commons.domain.RptId;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Generated;
