+++ src/test/java/it/pagopa/transactions/controllers/TransactionsControllerTest.java
@@ -1,8 +1,8 @@
 package it.pagopa.transactions.controllers;
 
 import io.vavr.control.Either;
-import it.pagopa.ecommerce.commons.domain.v1.PaymentToken;
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId;
+import it.pagopa.ecommerce.commons.domain.PaymentToken;
+import it.pagopa.ecommerce.commons.domain.TransactionId;
 import it.pagopa.ecommerce.commons.v1.TransactionTestUtils;
 import it.pagopa.generated.transactions.model.CtFaultBean;
 import it.pagopa.generated.transactions.server.model.*;
